MediaPlayer and VideoPlayer now do the same when it comes to volume control.
